document snapshots update + import document

+++ apihub-root/wallet/web-components/document/modals/document-snapshots-modal/document-snapshots-modal.js
@@ -1,5 +1,5 @@
 const documentModule = require("assistos").loadModule("document", {});
-const spaceModule = require("assistos").loadModule("space", {});
+import {formatTimestampToDate} from "../../../../utils/utils.js";
 export class DocumentSnapshotsModal{
     constructor(element, invalidate) {
         this.element = element;
@@ -9,20 +9,27 @@ export class DocumentSnapshotsModal{
         this.invalidate();
     }
     async beforeRender() {
+        let snapshots = this.document.snapshots;
+        if(this.document.type === documentModule.documentTypes.SNAPSHOT){
+            let documentInfo = JSON.parse(this.document.abstract);
+            snapshots = await documentModule.getDocumentSnapshots(assistOS.space.id, documentInfo.originalDocumentId);
+            snapshots.push({
+                documentId: documentInfo.originalDocumentId
+            })
+        }
         let snapshotsHTML = "";
         let headerHTML = `<div class="no-snapshots">no snapshots created</div>`;
-        if(this.document.snapshots.length > 0){
+        snapshots.sort((a, b) => b.timestamp - a.timestamp);
+        if(snapshots.length > 0){
             headerHTML = `<div class="list-header">
                                         <span class="snapshot-date">Date</span>
                                         <span class="snapshot-user">Created by</span>
-                                        <span class="export-snapshot">Export</span>
                                         <span>Delete</span>
                                     </div>`;
-            for (let snapshot of this.document.snapshots) {
-                snapshotsHTML += `<div class="document-snapshot">
-                                          <div class="snapshot-date">${snapshot.date}</div>
-                                          <div class="snapshot-user">${snapshot.userEmail}</div>
-                                          <img class="export-document pointer" data-local-action="exportDocument ${snapshot.id}" src="./wallet/assets/icons/download.svg" alt="download">
+            for (let snapshot of snapshots) {
+                snapshotsHTML += `<div class="document-snapshot" data-local-action="openSnapshot ${snapshot.documentId}">
+                                          <div class="snapshot-date">${formatTimestampToDate(snapshot.timestamp)}</div>
+                                          <div class="snapshot-user">${snapshot.email}</div>
                                           <img class="delete-snapshot pointer" data-local-action="deleteSnapshot ${snapshot.id}" src="./wallet/assets/icons/trash-can.svg" alt="delete">
                                         </div>`;
             }
@@ -30,7 +37,10 @@ export class DocumentSnapshotsModal{
         this.snapshotsHTML=`${headerHTML}${snapshotsHTML}`;
     }
     afterRender() {
-
+        let snapshotButton = this.element.querySelector(".snapshot-button");
+        if(this.document.type === documentModule.documentTypes.SNAPSHOT){
+            snapshotButton.classList.add("disabled");
+        }
     }
     closeModal() {
         assistOS.UI.closeModal(this.element);
@@ -40,13 +50,22 @@ export class DocumentSnapshotsModal{
             timestamp: Date.now(),
             email: assistOS.user.email
         }
-        await documentModule.addDocumentSnapshot(assistOS.space.id, this.document.id, snapshotData);
-        this.document.snapshots = await documentModule.getDocumentSnapshots(assistOS.space.id, this.document.id);
+        let snapshot = await documentModule.addDocumentSnapshot(assistOS.space.id, this.document.id, snapshotData);
+        this.document.snapshots.push(snapshot);
         this.invalidate();
     }
-    async deleteSnapshot(snapshotId){
+    async deleteSnapshot(targetElement, snapshotId){
+        let message = "Are you sure you want to delete this snapshot?";
+        let confirmation = await assistOS.UI.showModal("confirm-action-modal", {message}, true);
+        if (!confirmation) {
+            return;
+        }
         await documentModule.deleteDocumentSnapshot(assistOS.space.id, this.document.id, snapshotId);
-        this.document.snapshots.filter(snapshot => snapshot.id !== snapshotId);
+        this.document.snapshots = this.document.snapshots.filter(snapshot => snapshot.id !== snapshotId);
         this.invalidate();
+    }
+    openSnapshot(targetElement, documentId){
+        this.closeModal();
+        assistOS.UI.changeToDynamicPage("space-application-page", `${assistOS.space.id}/Space/document-view-page/${documentId}`);
     }
 }
